убрал логику из абстрактного сервиса, убрал лишние методы запроса к бд

+++ src/main/java/com/test/db/domain/ResultStat.java
@@ -12,6 +12,6 @@ public class ResultStat {
     private List<CustomerForStat> customers;
     private Integer totalDays;
     private Integer totalExpenses;
-    private BigDecimal avgExpenses ;
+    private int avgExpenses ;
 
 }
